refactor: Move bullet related things to the bullet module
